requirement.txt
windnd==1.0.7
pycryptodome==3.14.1
pyperclip==1.8.2